feature:红线/模板和git触发事件的一些调整 #1061
