+++ src/backend/ci/core/process/biz-process/build.gradle
@@ -44,7 +44,6 @@ dependencies {
     compile project(":core:artifactory:api-artifactory")
     compile project(":core:process:api-process")
     compile project(":core:quality:api-quality")
-    compile project(":core:quality:common-quality")
     compile project(":core:plugin:api-plugin")
     compile project(":core:plugin:codecc-plugin:common-codecc")
     compile project(":core:notify:api-notify")
