+++ src/backend/ci/settings.gradle
@@ -71,7 +71,6 @@ include ":core:misc:api-misc"
 include ":core:misc:biz-misc"
 include ":core:misc:boot-misc"
 include ":core:quality"
-include ":core:quality:common-quality"
 include ":core:quality:api-quality"
 include ":core:quality:biz-quality"
 include ":core:quality:boot-quality"
